src/requirements.txt
aiohttp==3.8.1
aioredis==2.0.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
charset-normalizer==2.0.12
frozenlist==1.3.0
hiredis==2.0.0
idna==3.3
msgpack==1.0.3
multidict==6.0.2
python-dotenv==0.19.2
PyYAML==6.0
tqdm==4.63.1
typing_extensions==4.1.1
websocket-client==1.3.1
websockets==10.2
yarl==1.7.2
